requirements.txt
Django>=4.2,<5.1
djangorestframework>=3.14,<3.16
mysqlclient
django-filter>=23.5
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27,<0.29
python-json-logger

